+++ src/formatters/expression.rs
@@ -1122,7 +1122,9 @@ fn format_hanging_expression_(
                 let formatted_expression = format_expression(ctx, expression, lhs_shape + 1); // 1 = opening parentheses
 
                 let expression_str = formatted_expression.to_string();
-                if !lhs_shape.add_width(2 + expression_str.len()).over_budget() {
+                if !contains_comments(expression)
+                    && !lhs_shape.add_width(2 + expression_str.len()).over_budget()
+                {
                     // The expression inside the parentheses is small, we do not need to break it down further
                     return Expression::Parentheses {
                         contained,